+++ package.json
@@ -9,7 +9,7 @@
         "prod": "npm run production",
         "production": "cross-env NODE_ENV=production node_modules/webpack/bin/webpack.js --progress --hide-modules --config=node_modules/laravel-mix/setup/webpack.config.js",
         "test": "ava",
-        "lint": "eslint --ext .js,.vue resources/js test/**/*.spec.js",
+        "lint": "eslint --ext .js,.vue resources/js/**/*.js",
         "lint:fix": "npm run lint -- --fix"
     },
     "devDependencies": {
@@ -56,10 +56,10 @@
     },
     "ava": {
         "require": [
-            "./test/helpers/setup.js"
+            "./resources/js/__tests__/helpers/setup.js"
         ],
         "files": [
-            "./test/**/*.spec.js"
+            "./resources/js/__tests__/**/*.spec.js"
         ],
         "sources": [
             "resources/js/**/*.js"
