+++ resources/js/__tests__/helpers/setup.js
@@ -1,0 +1,9 @@
+require('jsdom-global')()
+
+const hooks = require('require-extension-hooks')
+
+hooks('vue').plugin('vue').push()
+
+require('@babel/register')({
+  extensions: ['.vue', '.js']
+})
